Move common functions into a shared header files.
Below are changes that I have made:
1. Move common functions into a shared header file and update all examples.
2. Replace exit codes with enum values so it is more readable and consistent in all examples.
